requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
python-jose==3.3.0
sqlalchemy==2.0.25
python-decouple==3.8
